CartViewChanges
When clients scan/click a item, it can be added to the left view port now. It's functioned by CartControl.js.

Some inline style sheet was added in index.php

Embeded js code in scanBarcode.php so as to modify ProductsArray in localstorage each time when user scan a bar code.

The data is stored in terms of object in an array called selectedProductsArray, you can access it with the code below:
let existingProducts =  JSON.parse(myStorage.getItem('selectedProductArray'))